requirements.txt
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
groq>=0.4.0
anthropic>=0.18.0
pyperclip>=1.8.2
pyautogui>=0.9.54
python-dotenv>=1.0.0
